Add proctracer config option and other minor fixes
