+++ src/drakvuf.cpp
@@ -121,7 +121,8 @@ static gpointer timer(gpointer data)
     return NULL;
 }
 
-int drakvuf_c::start_plugins(const char *dump_folder)
+int drakvuf_c::start_plugins(const char *dump_folder,
+                             const char *proctracer_config)
 {
     int i, rc;
     for(i=0;i<__DRAKVUF_PLUGIN_LIST_MAX;i++)
@@ -132,6 +133,17 @@ int drakvuf_c::start_plugins(const char *dump_folder)
             struct filedelete_config c = {
                 .rekall_profile = this->rekall_profile,
                 .dump_folder = dump_folder
+            };
+
+            rc = this->plugins->start((drakvuf_plugin_t)i, &c);
+            break;
+        }
+
+        case PLUGIN_PROCTRACER:
+        {
+            struct proctracer_config c = {
+                .rekall_profile = this->rekall_profile,
+                .proctracer_config = proctracer_config
             };
 
             rc = this->plugins->start((drakvuf_plugin_t)i, &c);
