+++ src/drakvuf.h
@@ -142,7 +142,7 @@ class drakvuf_c {
         void pause();
         void resume();
         int inject_cmd(vmi_pid_t injection_pid, const char *inject_cmd);
-        int start_plugins(const char *dump_folder);
+        int start_plugins(const char *dump_folder, const char *proctracer_config);
 };
 
 #endif
