+++ src/plugins/plugins.h
@@ -117,6 +117,11 @@ struct filedelete_config {
     const char *dump_folder;
 };
 
+struct proctracer_config {
+    const char *rekall_profile;
+    const char *proctracer_config;
+};
+
 /***************************************************************************/
 
 typedef enum {
