+++ src/plugins/proctracer/proctracer.h
@@ -137,21 +137,13 @@ class proctracer: public plugin {
         };
 
         size_t *offsets;
- 
+
         output_format_t format;
         unordered_map<addr_t,list<mod_info*>> trace_status;
         unordered_map<string,list<addr_t>> mod_config;
 
-
         proctracer(drakvuf_t drakvuf, const void *config, output_format_t output);
         ~proctracer();
 };
-
-struct trace_info{
-    char* mod_name;
-    addr_t offset;
-    proctracer *p;
-};
-
 
 #endif
